improve 'hostname' compatibility
The original 'hostname' command provided by the project 'net-tools'
doesn't support the option '-I' (see last comment of the 'net-tools'
webpage).

Those tools invoking 'hostname' with the option '-I' (rootlesskit.sh
and cfssl.sh at the time of this writing) work on distros using a
version of 'hostname' supporting '-I' (basically Debian and RedHat
derivatives) out of the box. But it doesn't on those using the original
'hostname' (like ArchLinux and derivatives).

This patch adds support for ArchLinux and derivatives using the default
distro 'hostname' by falling back to using '-i' on those systems not
supporting the option '-I'. The only difference is that it depends on
name resolution (see [2] for the documentation of the option '-I').

[1] https://net-tools.sourceforge.io/
[2] https://linux.die.net/man/1/hostname

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>