bug: multi-node setup needs unique network names
In the case of using Podman (or a runtime that has a shared CNI
directory in the user home) and the case that the runtime
generates a cni file for each node network, if you have a
shared filesystem and a single, non-unique name, each node
will write a slightly different address in the CNI file
and clobber any previously written files (race condition).
This additional make multi-node command will replace
"default network" to be specific to the hostname and
avoid this.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>